requirements.txt
gradio
transformers
sentence-transformers
torch
torchvision 
langchain
langchain_openai
langchain-community
faiss-cpu
bitsandbytes
tiktoken
moviepy
opencv-python
boto3
numpy
accelerate
python-dotenv
plotly